requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-moment==1.0.5
wtforms==3.0.1
werkzeug==2.3.7
pysnmp==4.4.12
requests==2.31.0
pycryptodome==3.19.0
pyasn1==0.5.0
apscheduler==3.10.4
python-dotenv==1.0.0
bcrypt==4.0.1
matplotlib==3.8.0
pandas==2.1.1
sqlalchemy==2.0.21
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
markupsafe==2.1.3
scapy==2.5.0
flask-wtf==1.2.1
email-validator==2.0.0
